add transaction mapper unit tests

+++ app/src/main/java/com/christopherelias/blockchain/features/home/data_source/model/TransactionPerSecondResponse.kt
@@ -10,6 +10,6 @@ import com.squareup.moshi.Json
  */
 
 data class TransactionPerSecondResponse(
-    @field:Json(name = "x") val timeStamp: Double,
+    @field:Json(name = "x") val timeStamp: Long,
     @field:Json(name = "y") val transactionsPerSecondValue: Double
 )
